add OLED grating optimization pipeline with Bayesian RSM
- scripts/run_sweep.py: --exp namespace support, Ex source sweep (27 combos)
- scripts/recalc_efficiency.py: clip/z-only/raw P_total correction methods
- scripts/analyze_sweep.py: GP surrogate (RBF+WhiteKernel) contour/OFAT plots
- scripts/iterative_rsm.py: Sequential Bayesian optimization with EI acquisition
- main_feature/materials.py: AsymmetricSawtooth mask fix; add preset materials
- main_feature/CPML.py: E/H stagger coefficient correction
- runner.py: Poynting vector output, global NPZ saving, --exp routing
- CLAUDE.md: document light extraction pipeline, physical constraints

Result: Ex-source horizontal dipole, 37 simulations total,
best η_clip = 12.237% at p=38 h=26 d=0.40 (1.48× vs flat 8.241%)
GP optimum: p=37 h=26 d=0.35, η_pred=12.341%

+++ python-fdtd-gui-main/main_feature/materials.py
@@ -100,9 +100,9 @@ class Scene:
                 reshaped_mu   = {}
                 reshaped_cond = {}
 
-                eps= np.ones((self.Nx, self.Ny, self.Nz)) * self.eps0
-                mu= np.ones((self.Nx, self.Ny, self.Nz)) * self.mu0
-                cond = np.zeros((self.Nx, self.Ny, self.Nz))
+                eps  = np.ones ((self.Nx, self.Ny, self.Nz), dtype=np.float32) * self.eps0
+                mu   = np.ones ((self.Nx, self.Ny, self.Nz), dtype=np.float32) * self.mu0
+                cond = np.zeros((self.Nx, self.Ny, self.Nz), dtype=np.float32)
 
                 for shape, mat in self.objects:
                         mask = shape.mask(grid_shape)
